requirements.txt
python-dotenv
msal
requests
pandas
